refactor(deployment): extract service-token and wheel-build modules
Secret-minting recipes and dev-wheel manifest derivation move out of the
compose generator and deploy orchestrator into typed leaf modules
(service_tokens, wheel_build), re-exported from their original homes.
Also drops the dead dotted-name service lookup: services resolve by their
short name in the top-level services block, which the docstring now
states plainly.
