+++ src/osprey/deployment/compose_generator.py
@@ -16,6 +16,10 @@ from jinja2 import Environment, FileSystemLoader
 
 from osprey.deployment.facility_config import normalize_facility_config
 from osprey.deployment.runtime_helper import get_runtime_command, runtime_env
+from osprey.deployment.wheel_build import (
+    LOCAL_REQUIREMENTS_FILENAME,
+    _write_local_requirements_manifest,
+)
 from osprey.utils.config import ConfigBuilder
 from osprey.utils.log_filter import quiet_logger
 from osprey.utils.logger import get_logger
@@ -29,92 +33,38 @@ OUT_SRC_DIR = "repo_src"
 TEMPLATE_FILENAME = "docker-compose.yml.j2"
 COMPOSE_FILE_NAME = "docker-compose.yml"
 
-# Staged next to every dev wheel (see _copy_local_framework_for_override): the
-# wheel's own base dependency list, which the Dockerfiles' deps layer installs
-# with the build toolchain still present. Without it the deps layer primes only
-# the RELEASED PyPI pin's dependencies, and any base dep the local wheel adds
-# that the release lacks (e.g. a native sdist like softioc) would have to
-# compile in the later toolchain-less wheel-install layer — and fail.
-LOCAL_REQUIREMENTS_FILENAME = "osprey-local-requirements.txt"
-
 
 def find_service_config(config, service_name):
     """Locate service configuration and template path for deployment.
 
-    This function implements the service discovery logic for the container
-    management system, supporting both hierarchical service naming (full paths)
-    and legacy short names for backward compatibility. The system searches
-    through framework services and application-specific services to find
-    the requested service configuration.
-
-    Service naming supports three patterns:
-    1. Framework services: "osprey.service_name" or just "service_name"
-    2. Application services: "applications.app_name.service_name"
-    3. Legacy services: "service_name" (deprecated, for backward compatibility)
-
-    The function returns both the service configuration object and the path
-    to the Docker Compose template, enabling the caller to access service
-    settings and initiate template rendering.
+    Services are declared in the project config's top-level ``services:`` block,
+    keyed by service name — the same short name that appears in
+    ``deployed_services``. This function looks the name up in that block and
+    returns both the service configuration object and the path to its Docker
+    Compose template, so the caller can access service settings and initiate
+    template rendering.
 
     :param config: Configuration containing service definitions
     :type config: dict
-    :param service_name: Service identifier (short name or full dotted path)
+    :param service_name: Service identifier as it appears under ``services:``
+        (and in ``deployed_services``)
     :type service_name: str
-    :return: Tuple containing service configuration and template path,
-        or (None, None) if service not found
+    :return: Tuple containing service configuration and template path, or
+        ``(None, None)`` if the name is not declared under ``services:``
     :rtype: tuple[dict, str] or tuple[None, None]
 
     Examples:
-        Framework service discovery::
+        Service discovery::
 
-            >>> config = {'osprey': {'services': {'jupyter': {'path': 'services/osprey/jupyter'}}}}
-            >>> service_config, template_path = find_service_config(config, 'osprey.jupyter')
-            >>> print(template_path)  # 'services/osprey/jupyter/docker-compose.yml.j2'
-
-        Application service discovery::
-
-            >>> config = {'applications': {'als_assistant': {'services': {'mongo': {'path': 'services/applications/als_assistant/mongo'}}}}}
-            >>> service_config, template_path = find_service_config(config, 'applications.als_assistant.mongo')
-            >>> print(template_path)  # 'services/applications/als_assistant/mongo/docker-compose.yml.j2'
-
-        Legacy service discovery::
-
-            >>> config = {'services': {'legacy_service': {'path': 'services/legacy'}}}
-            >>> service_config, template_path = find_service_config(config, 'legacy_service')
-            >>> print(template_path)  # 'services/legacy/docker-compose.yml.j2'
-
-    .. note::
-       Legacy service support (services.* configuration) is deprecated and
-       will be removed in future versions. Use osprey.* or applications.*
-       naming patterns for new services.
+            >>> config = {'services': {'openobserve': {'path': 'services/openobserve'}}}
+            >>> service_config, template_path = find_service_config(config, 'openobserve')
+            >>> print(template_path)  # 'services/openobserve/docker-compose.yml.j2'
 
     .. seealso::
        :func:`setup_build_dir` : Processes discovered services for deployment
     """
-    # Handle full path notation (osprey.jupyter, applications.als_assistant.mongo)
-    if "." in service_name:
-        parts = service_name.split(".")
-
-        if parts[0] == "osprey" and len(parts) == 2:
-            # osprey.service_name
-            framework_services = config.get("osprey", {}).get("services", {})
-            service_config = framework_services.get(parts[1])
-            if service_config:
-                return service_config, os.path.join(service_config["path"], TEMPLATE_FILENAME)
-
-        elif parts[0] == "applications" and len(parts) == 3:
-            # applications.app_name.service_name
-            app_name, service_name_short = parts[1], parts[2]
-            applications = config.get("applications", {})
-            app_config = applications.get(app_name, {})
-            app_services = app_config.get("services", {})
-            service_config = app_services.get(service_name_short)
-            if service_config:
-                return service_config, os.path.join(service_config["path"], TEMPLATE_FILENAME)
-
-    # Handle short names - check legacy services for backward compatibility
-    legacy_services = config.get("services", {})
-    service_config = legacy_services.get(service_name)
+    services = config.get("services", {})
+    service_config = services.get(service_name)
     if service_config:
         return service_config, os.path.join(service_config["path"], TEMPLATE_FILENAME)
 
@@ -454,88 +404,6 @@ def _build_dev_wheel_cached(osprey_source_root):
 
     _wheel_build_cache[key] = cached_wheel
     return cached_wheel
-
-
-def _wheel_base_requirements(wheel_path):
-    """Extract the wheel's base (non-extra) ``Requires-Dist`` entries, sorted.
-
-    Parses ``*.dist-info/METADATA`` inside the wheel and returns every
-    ``Requires-Dist`` requirement string EXCEPT those gated behind an extra
-    (environment marker referencing ``extra``). Non-extra markers such as
-    ``python_version`` are kept verbatim so pip evaluates them in-container.
-
-    The result is fully deterministic for identical wheels — sorted, with
-    whitespace normalized — because the manifest built from it is COPY'd into
-    a BuildKit-content-hashed layer: byte-identical input keeps the deps-layer
-    cache warm across deploys.
-
-    :param wheel_path: Path to the built wheel file
-    :type wheel_path: str or pathlib.Path
-    :return: Sorted requirement strings (extras excluded)
-    :rtype: list[str]
-    :raises ValueError: If the wheel contains no ``*.dist-info/METADATA``
-    """
-    import zipfile
-    from email.parser import Parser
-
-    from packaging.requirements import Requirement
-
-    with zipfile.ZipFile(wheel_path) as whl:
-        metadata_names = sorted(
-            name for name in whl.namelist() if name.endswith(".dist-info/METADATA")
-        )
-        if not metadata_names:
-            raise ValueError(f"no *.dist-info/METADATA found in wheel {wheel_path}")
-        metadata = Parser().parsestr(whl.read(metadata_names[0]).decode("utf-8"))
-
-    requirements = []
-    for entry in metadata.get_all("Requires-Dist") or []:
-        req = Requirement(entry)
-        if req.marker is not None and "extra" in str(req.marker):
-            continue  # extra-gated dep: stays out of the base install set
-        # Collapse header-folding whitespace so the written line is stable
-        # and single-line; the requirement text itself stays verbatim.
-        requirements.append(" ".join(entry.split()))
-    return sorted(requirements)
-
-
-def _write_local_requirements_manifest(cached_wheel, out_dir):
-    """Write ``osprey-local-requirements.txt`` next to the staged dev wheel.
-
-    Content contract (shared with the service Dockerfiles, which COPY the file
-    and pip-install it in their toolchain-equipped deps layer): one requirement
-    per line, sorted, trailing newline — byte-identical for identical wheels.
-
-    :param cached_wheel: The cached local wheel the manifest derives from
-    :type cached_wheel: pathlib.Path
-    :param out_dir: Build context directory the wheel was staged into
-    :type out_dir: str
-    """
-    manifest_path = os.path.join(out_dir, LOCAL_REQUIREMENTS_FILENAME)
-    content = "".join(f"{line}\n" for line in _wheel_base_requirements(cached_wheel))
-    with open(manifest_path, "w", encoding="utf-8", newline="\n") as fh:
-        fh.write(content)
-
-
-def _staged_dev_artifact_paths(context_dir):
-    """All dev-staging artifacts (wheels + requirements manifest) in a context.
-
-    Callers that stage into a long-lived directory (project-image and persona
-    builds) snapshot this before/after staging so their ``finally`` cleanup
-    removes exactly what staging added — wheel AND manifest — and nothing that
-    was already there.
-
-    :param context_dir: Build context directory
-    :type context_dir: str
-    :return: Paths of staged dev artifacts currently present
-    :rtype: set[pathlib.Path]
-    """
-    root = Path(context_dir)
-    artifacts = set(root.glob("*.whl"))
-    manifest = root / LOCAL_REQUIREMENTS_FILENAME
-    if manifest.exists():
-        artifacts.add(manifest)
-    return artifacts
 
 
 def _copy_local_framework_for_override(out_dir):
@@ -1050,11 +918,12 @@ def setup_build_dir(template_path, config, container_cfg, dev_mode=False):
             logger.debug(f"Created flattened config.yml at {config_yml_dst}")
         except Exception as e:
             logger.warning(f"Failed to create flattened config: {e}")
-            # Fallback to copying original config — normalized, so a container
-            # never sees the raw `gitlab:` shape even on this degraded path. A
-            # config.yml the normalizer/YAML loader can't parse falls back
-            # further, to the previous verbatim copy, so this degraded path
-            # never blocks the build.
+            # Fallback to copying original config, normalized first. The primary
+            # deploy entry (`prepare_compose_files`) already rejects a removed
+            # `gitlab:` block up front, so this degraded path only ever sees a
+            # canonical config. A config.yml the normalizer/YAML loader can't
+            # parse falls back further, to a verbatim copy, so this degraded
+            # path never blocks the build.
             config_yml_src = "config.yml"
             if os.path.exists(config_yml_src):
                 config_yml_dst = os.path.join(out_dir, "config.yml")
